refactor: BaseEntity package 구조 변경
