+++ backend/src/main/java/com/allog/dallog/schedule/domain/Schedule.java
@@ -1,6 +1,6 @@
 package com.allog.dallog.schedule.domain;
 
-import com.allog.dallog.common.BaseEntity;
+import com.allog.dallog.global.entity.BaseEntity;
 import com.allog.dallog.category.domain.Category;
 import com.allog.dallog.schedule.exception.InvalidScheduleException;
 import java.time.LocalDateTime;
