+++ backend/src/main/java/com/allog/dallog/subscription/domain/Subscription.java
@@ -1,7 +1,7 @@
 package com.allog.dallog.subscription.domain;
 
 import com.allog.dallog.auth.exception.NoPermissionException;
-import com.allog.dallog.common.BaseEntity;
+import com.allog.dallog.global.entity.BaseEntity;
 import com.allog.dallog.category.domain.Category;
 import com.allog.dallog.member.domain.Member;
 import javax.persistence.Column;
